### 2025-02-20 #7808 @editorialbot
[PRE REVIEW]: PyBispectra: A toolbox for advanced electrophysiological signal processing using the bispectrum
**Submitting author:** @tsbinns (<a href="http://orcid.org/0000-0003-0657-0891">Thomas Samuel Binns</a>)
**Repository:** https://github.com/braindatalab/PyBispectra
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.2.1
