+++ README.md
@@ -13,6 +13,9 @@ More information on the [installation](https://pybispectra.readthedocs.io/en/mai
 ## Use:
 To get started with the toolbox, check out the [documentation](https://pybispectra.readthedocs.io/en/main/) and [examples](https://pybispectra.readthedocs.io/en/main/examples.html).
 
+## Contributing & Development:
+If you encounter issues with the package, want to suggest improvements, or have made any changes which you would like to see officially supported, please refer to the [development](https://pybispectra.readthedocs.io/en/main/development.html) page. A unit test suite is included and must be expanded where necessary to validate any changes.
+
 ## Citing:
 If you use this toolbox in your work, please include the following citation:<br/>
 Binns, T. S., Pellegrini, F., Jurhar, T., & Haufe, S. (2023). PyBispectra. DOI: [10.5281/zenodo.8377820](https://doi.org/10.5281/zenodo.8377820)
