+++ CONTRIBUTING.md
@@ -1,0 +1,5 @@
+# Contributing to PyBispectra
+
+All contributions are welcomed, whether you wish to report issues you have encountered, have suggestions for improvements, or have made changes which you would like to see officially added to the package. For more information, see the [development](https://pybispectra.readthedocs.io/en/main/development.html) page.
+
+Users and contributors to PyBispectra are expected to follow our [Code of Conduct](CODE_OF_CONDUCT.md).
