### 2025-02-20 79439ee
[PAPER] Update contribution guidelines and add code of conduct (#55)
