+++ docs/source/development.rst
@@ -26,7 +26,7 @@ Contributing to PyBispectra
 ---------------------------
 
     This project and everyone participating in it is governed by our `Code of Conduct
-    <https://github.com/braindatalab/PyBispectra/CODE_OF_CONDUCT.md>`_. By
+    <https://github.com/braindatalab/PyBispectra/blob/main/CODE_OF_CONDUCT.md>`_. By
     participating, you are expected to uphold this code.
 
 If you encounter any issues with the package or wish to suggest improvements, please
