### 2025-02-25 5e73fb1
[pre-commit.ci] pre-commit autoupdate (#57)
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
