+++ paper/paper.bib
@@ -134,6 +134,14 @@
   doi={10.1073/pnas.1214546110}
 }
 
+@article{JurharPreprint,
+  title={{Estimating Time Delays between Signals under Mixed Noise Influence with Novel Cross- and Bispectral Methods}}, 
+  author={Tin Jurhar and Franziska Pellegrini and Ana I. Nuñes del Toro and Tilman Stephani and Guido Nolte and Stefan Haufe},
+  year={2025},
+  journal={arXiv},
+  doi={10.48550/arXiv.2502.17474}
+}
+
 @inproceedings{Lam2015,
   title={Numba: A {LLVM}-based {P}ython {JIT} compiler},
   author={Lam, Siu Kwan and Pitrou, Antoine and Seibert, Stanley},
