+++ docs/source/refs.bib
@@ -102,11 +102,12 @@
   doi={10.1073/pnas.1214546110}
 }
 
-@article{JurharInPrep,
-  title={Estimating signal time-delays under mixed noise influence with novel cross- and bispectrum methods},
-  author={Jurhar, Tin and others},
-  journal={In Preparation},
-  year={In Preparation}
+@article{JurharPreprint,
+  title={{Estimating Time Delays between Signals under Mixed Noise Influence with Novel Cross- and Bispectral Methods}}, 
+  author={Tin Jurhar and Franziska Pellegrini and Ana I. Nuñes del Toro and Tilman Stephani and Guido Nolte and Stefan Haufe},
+  year={2025},
+  journal={arXiv},
+  doi={10.48550/arXiv.2502.17474}
 }
 
 @article{Kovach2018,
