### 2025-03-17 6f5af2f
[pre-commit.ci] pre-commit autoupdate (#61)
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
