### 2025-03-21 #7808 @britta-wstnr
Hi @Kevin-Mattheus-Moerman, @tsbinns is a regular contributor to MNE-Python, which I am maintainer and steering council member of. With that, I feel like this is a bit too close to home and a potential COI?
