### 2025-04-15 1c66b04
[pre-commit.ci] pre-commit autoupdate (#64)
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
