### 2025-04-22 33215cf
[pre-commit.ci] pre-commit autoupdate (#65)
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
