### 2025-04-29 972d5e1
[pre-commit.ci] pre-commit autoupdate (#66)
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
