### 2025-05-26 f8e0189
[pre-commit.ci] pre-commit autoupdate (#70)
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
