### 2025-06-03 3e2db27
[pre-commit.ci] pre-commit autoupdate (#71)
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
