### 2025-06-21 #7808 @zhubonan
Hi @ntolley, @richrobe and  @saran-nns would you be able to help review this JOSS submission?

The review process of JOSS is unique as it happens on GitHub. The process is open and transparent and interaction between the reviewers and the author are done via GitHub issues (and encouraged). You can learn about the reviewing guideline here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html.
