### 2025-07-08 8dac153
[pre-commit.ci] pre-commit autoupdate (#75)
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
