### 2025-07-13 #8504 @ntolley
Hi @tsbinns,

This is a really interesting package! I definitely think I could find some use for this in my own work. I have a couple minor suggestions for the JOSS paper that I think would strengthen the writeup. On the whole everything was quite clean and well explained

1) I think there are some packages that are worth mentioning to highlight what makes PyBispectrum unique. Specifically NeuroDSP (https://neurodsp-tools.github.io/neurodsp/index.html), which has a few overlapping utility functions for phase analysis, but does not approach it from the bispectrum. There is also Elephant (https://elephant.readthedocs.io/en/latest/index.html) which is a bit more catered to spike/LFP analysis, so I’ll leave it to your discretion if it’s worth citing this one. The only thing to emphasize is that these are packages that provide spectral analysis specifically for ephys data (besides MNE), but your package is unique by using the bispectrum.

2) In the figure 1 caption, I think the 3rd sentence could state “Tools are **also** provided for computing cross-frequency…”

3) On the github README, I think you could add 1-2 sentences that pull from the JOSS paper summary/statement of need to make very explicit who the target audience is, and why the bispectrum is useful. Just something to the nature of “PAC, TDE, etc. are useful for neuroscientists because of <>. This package provides these analyses using the bispectrum which overcomes limitations of conventional approaches to these analysis methods." Some of this is in the motivation of the docs, but it could be briefly highlighted on the first page.
