### 2025-07-15 6059bd4
[pre-commit.ci] pre-commit autoupdate (#76)
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
