### 2025-07-27 006dcfa
Revert "Update from review"
This reverts commit 2d70652421bfde1279f9085681dd72f45032734c.
