+++ paper/paper.md
@@ -52,9 +52,9 @@ Various forms of information can be extracted from neural time series data. Of t
 
 Analysis of phase-amplitude coupling, time delays, and non-sinusoidal waveshape provide important insights into interneuronal communication [@Canolty2010;@Silchenko2010;@Sherman2016]. Studies of these features in neural data have been used to investigate core functions such as movement and memory, including their perturbation in disease [@deHemptinne2013;@Cole2017;@Bazzigaluppi2018;@Binns2024]. However, traditional analysis methods have critical limitations that hinder their utility. In contrast, the bispectrum - the Fourier transform of the third order moment [@Nikias1987] - can be used for the analysis of phase-amplitude coupling [@Zandvoort2021], non-sinusoidal waveshape [@Bartz2019], and time delays [@Nikias1988], overcoming many traditional limitations.
 
-Despite these benefits, the bispectrum has seen little use in neuroscience research, in part due to the lack of an accessible toolbox tailored to electrophysiology data. Code written in MATLAB exists for some analyses (see e.g., [github.com/sccn/roiconnect](https://github.com/sccn/roiconnect), [github.com/ZuseDre1/AnalyzingWaveshapeWithBicoherence](https://github.com/ZuseDre1/AnalyzingWaveshapeWithBicoherence)), however it is spread across multiple repositories and often not as toolboxes. Furthermore, this requires a paid MATLAB license, limiting its accessibility. Code for computing the bispectrum exists in the free-to-use Python language [@Bachetti2024], however these implementations are not tailored for electrophysiology data, and while other Python packages perform some of these analyses on electrophysiology data [@Cole2019;@Denker2024], they do not make use of the bispectrum. The `PyBispectra` package addresses this by providing a comprehensive toolbox for bispectral analysis of electrophysiology data (\autoref{fig:overview}), including tutorials to facilitate an understanding of these analyses.
+Despite these benefits, the bispectrum has seen little use in neuroscience research, in part due to the lack of an accessible toolbox tailored to electrophysiology data. Code written in MATLAB exists for some analyses (see e.g., [github.com/sccn/roiconnect](https://github.com/sccn/roiconnect), [github.com/ZuseDre1/AnalyzingWaveshapeWithBicoherence](https://github.com/ZuseDre1/AnalyzingWaveshapeWithBicoherence)), however it is spread across multiple repositories and often not as toolboxes. Furthermore, this requires a paid MATLAB license, limiting its accessibility. Code for computing the bispectrum exists in the free-to-use Python language - e.g., @Bachetti2024 - however these implementations are not tailored for electrophysiology data. The `PyBispectra` package addresses these problems by providing a comprehensive toolbox for bispectral analysis of electrophysiology data (\autoref{fig:overview}), including tutorials to facilitate an understanding of these analyses.
 
-![\label{fig:overview}Overview of the `PyBispectra` toolbox. Optional preprocessing methods are supported for the multivariate analysis of waveshape. Tools are provided for computing spectral representations of time series data, as well as for computing cross-frequency coupling, time delays, and non-sinusoidal waveshape, with schematic visualisations of results shown. Also shown is an example code snippet for analysing phase-amplitude coupling.](Overview.svg)
+![\label{fig:overview}Overview of the `PyBispectra` toolbox. Optional preprocessing methods are supported for the multivariate analysis of waveshape. Tools are provided for computing spectral representations of time series data. Tools are provided for computing cross-frequency coupling, time delays, and non-sinusoidal waveshape, with schematic visualisations of results shown. Also shown is an example code snippet for analysing phase-amplitude coupling.](Overview.svg)
 
 # Features
 
