### 2025-07-27 9051d4e
Reapply "Update from review"
This reverts commit 006dcfacde6e82b0dfb3d0954ef72c777de2fa63.
