### 2025-07-27 2d098db
Revert "Reapply "Update from review""
This reverts commit 9051d4e8bd4ed5ded48cf76e4e4e365b51adc116.
