### 2025-07-27 c59f1be
Reapply "Update from review"
This reverts commit 006dcfacde6e82b0dfb3d0954ef72c777de2fa63.
