+++ paper/paper.bib
@@ -101,6 +101,26 @@
   doi={10.1523/JNEUROSCI.2208-16.2017}
 }
 
+@article{Cole2019,
+  title={{NeuroDSP: A package for neural digital signal processing}},
+  author={Cole, Scott and Donoghue, Thomas and Gao, Richard and Voytek, Bradley}, 
+  journal={Journal of Open Source Software},
+  volume={4},
+  number={36},
+  pages={1272},
+  year={2019},
+  publisher={The Open Journal},
+  doi={10.21105/joss.01272},
+}
+
+@software{Denker2024,
+  title={Elephant},
+  author={Denker, Michael and Kern, Moritz and Richter, Felician},
+  year={2024},
+  publisher={Zenodo},
+  doi={10.5281/zenodo.13133971},
+}
+
 @article{Giehl2021,
   title={Dissociating harmonic and non-harmonic phase-amplitude coupling in the human brain},
   author={Giehl, Janet and Noury, Nima and Siegel, Markus},
