### 2025-07-27 7b4d479
[PAPER] Avoid name disambiguation in citations (#81)
