+++ paper/paper.bib
@@ -91,7 +91,7 @@
 
 @article{Cole2017,
   title={Nonsinusoidal beta oscillations reflect cortical pathophysiology in {P}arkinson's disease},
-  author={Cole, Scott R. and van der Meij, Roemer and Peterson, Erik J. and de Hemptinne, Coralie and Starr, Philip A. and Voytek, Bradley},
+  author={Cole, Scott and van der Meij, Roemer and Peterson, Erik J. and de Hemptinne, Coralie and Starr, Philip A. and Voytek, Bradley},
   journal={Journal of Neuroscience},
   volume={37},
   number={18},
