### 2025-07-27 27380c6
[PAPER] Update citations
