+++ paper/paper.md
@@ -50,7 +50,7 @@ Various forms of information can be extracted from neural time series data. Of t
 
 # Statement of need
 
-Analysis of phase-amplitude coupling, time delays, and non-sinusoidal waveshape provide important insights into interneuronal communication [@Canolty2010;@Silchenko2010;@Sherman2016]. Studies of these features in neural data have been used to investigate core functions such as movement and memory, including their perturbation in disease [@deHemptinne2013;@Cole2017;@Bazzigaluppi2018;@Binns2024]. However, traditional analysis methods have critical limitations that hinder their utility. In contrast, the bispectrum - the Fourier transform of the third order moment [@Nikias1987] - can be used for the analysis of phase-amplitude coupling [@Zandvoort2021], non-sinusoidal waveshape [@Bartz2019], and time delays [@Nikias1988], overcoming many traditional limitations.
+Analysis of phase-amplitude coupling, time delays, and non-sinusoidal waveshape provide important insights into interneuronal communication [@Canolty2010;@Silchenko2010;@Sherman2016]. Studies of these features in neural data have been used to investigate core functions such as movement and memory, including their perturbation in disease [@deHemptinne2013;@Cole2017;@Bazzigaluppi2018;@Binns2025]. However, traditional analysis methods have critical limitations that hinder their utility. In contrast, the bispectrum - the Fourier transform of the third order moment [@Nikias1987] - can be used for the analysis of phase-amplitude coupling [@Zandvoort2021], non-sinusoidal waveshape [@Bartz2019], and time delays [@Nikias1988], overcoming many traditional limitations.
 
 Despite these benefits, the bispectrum has seen little use in neuroscience research, in part due to the lack of an accessible toolbox tailored to electrophysiology data. Code written in MATLAB exists for some analyses (see e.g., [github.com/sccn/roiconnect](https://github.com/sccn/roiconnect), [github.com/ZuseDre1/AnalyzingWaveshapeWithBicoherence](https://github.com/ZuseDre1/AnalyzingWaveshapeWithBicoherence)), however it is spread across multiple repositories and often not as toolboxes. Furthermore, this requires a paid MATLAB license, limiting its accessibility. Code for computing the bispectrum exists in the free-to-use Python language [@Bachetti2024], however these implementations are not tailored for electrophysiology data, and while other Python packages perform some of these analyses on electrophysiology data [@Cole2019;@Denker2024], they do not make use of the bispectrum. The `PyBispectra` package addresses this by providing a comprehensive toolbox for bispectral analysis of electrophysiology data (\autoref{fig:overview}), including tutorials to facilitate an understanding of these analyses.
 
@@ -64,7 +64,7 @@ Phase-amplitude coupling is the interaction between the phase of a lower frequen
 
 ## Time delays
 
-Time delay analysis identifies latencies of information transfer between signals, providing  insight into the physical connections between brain regions [@Silchenko2010;@Binns2024]. A traditional analysis method is cross-correlation, quantifying the similarity of signals at a set of time lags. However, this approach has a limited robustness to noise [@Nikias1988] and a vulnerability to spurious zero time lag interactions arising due to volume conduction and source mixing in the sensor space. On the other hand, the bispectrum is resilient to Gaussian noise [@Nikias1988], and antisymmetrisation can be used to correct for spurious zero time lag interactions [@Chella2014;@JurharPreprint]. `PyBispectra` provides tools for bispectral time delay analysis, with options for antisymmetrisation.
[…]
