### 2025-07-27 596f8a4
[PAPER] Update citations (#82)
