+++ paper/paper.bib
@@ -33,20 +33,22 @@
   doi={10.1111/jnc.14136}
 }
 
-@article{Binns2024,
+@article{Binns2025,
   title={Shared pathway-specific network mechanisms of dopamine and deep brain stimulation for the treatment of {P}arkinson's disease},
   author={Binns, Thomas S. and K{\"o}hler, Richard M. and Vanhoecke, Jojo and Chikermane, Meera and Gerster, Moritz and Merk, Timon and Pellegrini, Franziska and Busch, Johannes L. and Habets, Jeroen G.V. and Cavallo, Allesia and Beyer, Jean-Christin and Al-Fatly, Bassam and Li, Ningfei and Horn, Andreas and Krause, Patricia and Faust, Katharina and Schneider, Gerd-Helge and Haufe, Stefan and K{\"u}hn, Andrea A. and Neumann, Wolf-Julian},
-  journal={bioRxiv},
-  pages={2024--04},
-  year={2024},
-  publisher={Cold Spring Harbor Laboratory},
-  doi={10.1101/2024.04.14.586969}
+  journal={Nature Communications},
+  volume={16},
+  number={1},
+  pages={3587},
+  year={2025},
+  publisher={Nature Publishing Group},
+  doi={10.1038/s41467-025-58825-z}
 }
 
 @article{Canolty2006,
   title={High gamma power is phase-locked to theta oscillations in human neocortex},
   author={Canolty, Ryan T. and Edwards, Erik and Dalal, Sarang S. and Soltani, Maryam and Nagarajan, Srikantan S. and Kirsch, Heidi E. and Berger, Mitchel S. and Barbaro, Nicholas M. and Knight, Robert T.},
-  journal={science},
+  journal={Science},
   volume={313},
   number={5793},
   pages={1626--1628},
