### 2025-07-27 #8504 @tsbinns
Hi @ntolley, many thanks for the constructive review! I have incorporated the suggested changes [here](https://github.com/braindatalab/PyBispectra/pull/80). Please let me know if I should address these points further.
