+++ src/pybispectra/general/general.py
@@ -145,8 +145,11 @@ class Bispectrum(_General):
         :math:`f_1` and :math:`f_2` correspond to a lower and higher frequency,
         respectively; and :math:`<>` represents the average value over epochs.
 
-        The bispectrum is computed between all values of ``f1s`` and ``f2s``. If any
-        value of ``f1s`` is higher than ``f2s``, a :obj:`numpy.nan` value is returned.
+        The bispectrum is computed between all values of ``f1s`` and ``f2s``.
+
+        .. warning::
+            For values of ``f1s`` higher than ``f2s`` or where ``f2s + f1s`` exceeds the
+            Nyquist frequency, a :obj:`numpy.nan` value is returned.
         """
         self._reset_attrs()
 
@@ -308,8 +311,11 @@ class Threenorm(_General):
         :math:`f_1` and :math:`f_2` correspond to a lower and higher frequency,
         respectively; and :math:`<>` represents the average value over epochs.
 
-        The threenorm is computed between all values of ``f1s`` and ``f2s``. If any
-        value of ``f1s`` is higher than ``f2s``, a :obj:`numpy.nan` value is returned.
+        The threenorm is computed between all values of ``f1s`` and ``f2s``.
+
+        .. warning::
+            For values of ``f1s`` higher than ``f2s`` or where ``f2s + f1s`` exceeds the
+            Nyquist frequency, a :obj:`numpy.nan` value is returned.
 
         References
         ----------
