+++ src/pybispectra/cfc/aac.py
@@ -94,8 +94,11 @@ class AAC(_ProcessFreqBase):
         frequency in each epoch, with coupling being averaged across epochs
         :footcite:`Giehl2021`.
 
-        AAC is computed between all values of ``f1s`` and ``f2s``. If any value of
-        ``f1s`` is higher than ``f2s``, a :obj:`numpy.nan` value is returned.
+        AAC is computed between all values of ``f1s`` and ``f2s``.
+        
+        .. warning::
+            For values of ``f1s`` higher than ``f2s`` or where ``f2s + f1s`` exceeds the
+            Nyquist frequency, a :obj:`numpy.nan` value is returned.
 
         References
         ----------
