### 2025-07-27 5a66ab7
Remove redundant examples
