### 2025-07-28 c117b48
Add minimal core dependency requirements
