+++ pyproject.toml
@@ -10,13 +10,13 @@ classifiers = [
   "Programming Language :: Python :: 3",
 ]
 dependencies = [
-  "joblib",
-  "matplotlib",
-  "mne>1.6",
-  "numba",
-  "numpy",
-  "scikit-learn",
-  "scipy",
+  "joblib>=1.0.0",
+  "matplotlib>=3.5.0",
+  "mne>=1.7",
+  "numba>=0.55.0",
+  "numpy>=1.21.2",
+  "scikit-learn>=1.0",
+  "scipy>=1.7.1",
 ]
 description = "A Python signal processing package for computing spectral-domain and time-domain interactions using the bispectrum."
 name = "pybispectra"
