### 2025-07-28 1945159
Update installation instructions
