+++ src/pybispectra/cfc/__init__.py
@@ -1,7 +1,5 @@
 """Tools for handling cross-frequency coupling analysis."""
 
-__version__ = "1.2.1dev"
-
 from .aac import AAC
 from .pac import PAC
 from .ppc import PPC
