+++ src/pybispectra/waveshape/__init__.py
@@ -1,5 +1,3 @@
 """Tools for handling waveshape analysis."""
 
-__version__ = "1.2.1dev"
-
 from .waveshape import WaveShape
