+++ src/pybispectra/data/__init__.py
@@ -1,5 +1,3 @@
 """Storage and tools for handling data files."""
 
-__version__ = "1.2.1dev"
-
 from .example_data import DATASETS, get_example_data_paths
