### 2025-07-29 847acaa
Add pooch as optional dependency
