+++ pyproject.toml
@@ -32,6 +32,7 @@ doc = [
   "ipywidgets",
   "notebook",
   "numpydoc",
+  "pooch>=1.3.0",
   "pydata-sphinx-theme",
   "sphinx",
   "sphinx-copybutton",
@@ -49,7 +50,7 @@ lint = [
   "toml-sort",
   "yamllint",
 ]
-test = ["coverage", "pytest"]
+test = ["coverage", "pooch>=1.3.0", "pytest"]
 
 [project.urls]
 "Bug Tracker" = "https://github.com/braindatalab/PyBispectra/issues"
