### 2025-07-29 5ca3845
Migrate test data handling to pooch
