+++ src/pybispectra/utils/__init__.py
@@ -2,4 +2,11 @@
 
 from .ged import SpatioSpectralFilter
 from .results import ResultsCFC, ResultsGeneral, ResultsTDE, ResultsWaveShape
-from .utils import compute_fft, compute_rank, compute_tfr, set_precision
+from .utils import (
+    compute_fft,
+    compute_rank,
+    compute_tfr,
+    get_example_data_paths,
+    set_precision,
+    DATASETS,
+)
