+++ src/pybispectra/data/__init__.py
@@ -1,3 +1,0 @@
-"""Storage and tools for handling data files."""
-
-from .example_data import DATASETS, get_example_data_paths

[…]
