### 2025-07-29 be01b08
Fix errors in pooch migration
