### 2025-07-29 ed6085b
Download data before starting tests
