### 2025-07-29 e0922fe
Update pooch data downloading
