+++ .github/workflows/unit_tests.yml
@@ -59,7 +59,7 @@ jobs:
       - run: pip install -e .
       - run: python -c "import pybispectra; import mne; print(f'PyBispectra {pybispectra.__version__}\n'); mne.sys_info()"
       - name: Download testing data
-        run: python -c "from pybispectra.utils import DATASETS, get_example_data_paths; (get_example_data_paths(data, verbose=True) for data in DATASETS.keys())"
+        run: python -c "from pybispectra.utils import DATASETS, get_example_data_paths; (get_example_data_paths(data) for data in DATASETS.keys())"
       - name: Run pytest
         run: python -m coverage run && coverage report
 
@@ -91,6 +91,6 @@ jobs:
       - name: Display versions and environment information
         run: python -c "import pybispectra; import mne; print(f'PyBispectra {pybispectra.__version__}\n'); mne.sys_info()"
       - name: Download testing data
-        run: python -c "from pybispectra.utils import DATASETS, get_example_data_paths; (get_example_data_paths(data, verbose=True) for data in DATASETS.keys())"
+        run: python -c "from pybispectra.utils import DATASETS, get_example_data_paths; (get_example_data_paths(data) for data in DATASETS.keys())"
       - name: Run pytest
         run: python -m coverage run && coverage report
