+++ src/pybispectra/utils/__init__.py
@@ -3,10 +3,10 @@
 from .ged import SpatioSpectralFilter
 from .results import ResultsCFC, ResultsGeneral, ResultsTDE, ResultsWaveShape
 from .utils import (
+    DATASETS,
     compute_fft,
     compute_rank,
     compute_tfr,
     get_example_data_paths,
     set_precision,
-    DATASETS,
 )
