### 2025-07-29 #8504 @tsbinns
Hi @richrobe, thank you very much for the detailed feedback. I have incorporated the suggested changes [here](https://github.com/braindatalab/PyBispectra/pull/83). In particular:

- I have updated the introduction of the paper to better introduce the methods in question and their applicability, including a reframing to better describe the eligible forms of electrophysiology data.

- Greatly reducing the package size by moving the example data outside of the source directory and using `pooch` to handle its acquisition.

- Offering additional instructions for installing the package through `uv`.

I have also made updates based on your other comments. Please let me know if I should address these points further.
