+++ src/pybispectra/__init__.py
@@ -1,9 +1,8 @@
 """Initialisation of the PyBispectra package."""
 
-__version__ = "1.2.1dev"
+__version__ = "1.3.0+dev"
 
 from .cfc import AAC, PAC, PPC
-from .data import get_example_data_paths
 from .general import Bispectrum, Threenorm
 from .tde import TDE
 from .utils import (
@@ -15,6 +14,7 @@ from .utils import (
     compute_fft,
     compute_rank,
     compute_tfr,
+    get_example_data_paths,
     set_precision,
 )
 from .waveshape import WaveShape
