+++ tests/test_general.py
@@ -5,9 +5,8 @@ import pytest
 from numpy.random import RandomState
 
 from pybispectra.cfc import PAC
-from pybispectra.data import get_example_data_paths
 from pybispectra.general import Bispectrum, Threenorm
-from pybispectra.utils import ResultsGeneral, compute_fft
+from pybispectra.utils import ResultsGeneral, compute_fft, get_example_data_paths
 from pybispectra.utils._utils import _generate_data
 from pybispectra.waveshape import WaveShape
 
