### 2025-07-30 2d829b9
[PAPER] Update README for paper reframing (#87)
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
