### 2025-08-12 08a9c37
[pre-commit.ci] pre-commit autoupdate (#90)
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
