### 2025-08-14 b2e3967
[DOC] Update f2+f1 boundary warnings (#78)
Co-authored-by: Thomas S. Binns <t.s.binns@outlook.com>
