+++ paper/paper.bib
@@ -1,288 +1,0 @@
-@article{Bachetti2024,
-  title={Stingray 2: A fast and modern {P}ython library for spectral timing},
-  author={Matteo Bachetti and Daniela Huppenkothen and Abigail Stevens and John Swinbank and Guglielmo Mastroserio and Matteo Lucchini and Eleonora Veronica Lai and Johannes Buchner and Amogh Desai and Gaurav Joshi and Francesco Pisanu and Sri Guru Datta Pisupati and Swapnil Sharma and Mihir Tripathi and Dhruv Vats},
-  journal={Journal of Open Source Software},
-  volume={9},
-  number={102},
-  pages={7389},
-  year={2024},
-  publisher={The Open Journal},
-  doi={10.21105/joss.07389}
-}
-
-@article{Bartz2019,
-  title={Analyzing the waveshape of brain oscillations with bicoherence},
-  author={Bartz, Sarah and Avarvand, Forooz Shahbazi and Leicht, Gregor and Nolte, Guido},
-  journal={NeuroImage},
-  volume={188},
-  pages={145--160},
-  year={2019},
-  publisher={Elsevier},
-  doi={10.1016/j.neuroimage.2018.11.045}
-}
-
-@article{Bazzigaluppi2018,
-  title={Early-stage attenuation of phase-amplitude coupling in the hippocampus and medial prefrontal cortex in a transgenic rat model of {A}lzheimer's disease},
-  author={Bazzigaluppi, Paolo and Beckett, Tina L. and Koletar, Margaret M. and Lai, Aaron Y. and Joo, Illsung L. and Brown, Mary E. and Carlen, Peter L. and McLaurin, JoAnne and Stefanovic, Bojana},
-  journal={Journal of Neurochemistry},
-  volume={144},
-  number={5},
-  pages={669--679},
-  year={2018},
-  publisher={Wiley Online Library},
-  doi={10.1111/jnc.14136}
-}
-
-@article{Binns2025,
-  title={Shared pathway-specific network mechanisms of dopamine and deep brain stimulation for the treatment of {P}arkinson's disease},
-  author={Binns, Thomas S. and K{\"o}hler, Richard M. and Vanhoecke, Jojo and Chikermane, Meera and Gerster, Moritz and Merk, Timon and Pellegrini, Franziska and Busch, Johannes L. and Habets, Jeroen G.V. and Cavallo, Allesia and Beyer, Jean-Christin and Al-Fatly, Bassam and Li, Ningfei and Horn, Andreas and Krause, Patricia and Faust, Katharina and Schneider, Gerd-Helge and Haufe, Stefan and K{\"u}hn, Andrea A. and Neumann, Wolf-Julian},
-  journal={Nature Communications},
-  volume={16},
-  number={1},
-  pages={3587},
-  year={2025},
-  publisher={Nature Publishing Group},
-  doi={10.1038/s41467-025-58825-z}
-}
-
-@article{Canolty2006,
-  title={High gamma power is phase-locked to theta oscillations in human neocortex},
-  author={Canolty, Ryan T. and Edwards, Erik and Dalal, Sarang S. and Soltani, Maryam and Nagarajan, Srikantan S. and Kirsch, Heidi E. and Berger, Mitchel S. and Barbaro, Nicholas M. and Knight, Robert T.},
-  journal={Science},
-  volume={313},
-  number={5793},
-  pages={1626--1628},
-  year={2006},
-  publisher={American Association for the Advancement of Science},
-  doi={10.1126/science.1128115}
-}
-
-@article{Canolty2010,
-  title={The functional role of cross-frequency coupling},
[…]
