### 2025-08-15 e2a333d
[MAINT] Prepare for v1.2.2 release (#91)
