+++ paper/paper.md
@@ -1,87 +1,0 @@
----
-title: 'PyBispectra: A toolbox for advanced electrophysiological signal processing using the bispectrum'
-tags:
-  - Python
-  - neuroscience
-  - signal processing
-  - bispectrum
-authors:
-  - name: Thomas S. Binns
-    orcid: 0000-0003-0657-0891
-    affiliation: "1, 2, 3"
-    corresponding: true
-  - name: Franziska Pellegrini
-    orcid: 0000-0001-9769-1597
-    affiliation: "3, 4"
-  - name: Tin Jurhar
-    orcid: 0000-0002-8804-2349
-    affiliation: "5, 6"
-  - name: Tien D. Nguyen
-    orcid: 0009-0008-9867-0964
-    affiliation: "3, 4"
-  - name: Richard M. Köhler
-    orcid: 0000-0002-5219-1289
-    affiliation: 1
-  - name: Stefan Haufe
-    orcid: 0000-0003-1470-9195
-    affiliation: "2, 3, 4, 5, 7"
-affiliations:
- - name: Movement Disorders Unit, Charité - Universitätsmedizin Berlin, Germany
-   index: 1
- - name: Einstein Center for Neurosciences Berlin, Charité - Universitätsmedizin Berlin, Germany
-   index: 2
- - name: Bernstein Center for Computational Neuroscience Berlin, Germany
-   index: 3
- - name: Berlin Center for Advanced Neuroimaging, Charité - Universitätsmedizin Berlin, Germany
-   index: 4
- - name: Electrical Engineering and Computer Science Department, Technische Universität Berlin, Germany
-   index: 5
- - name: Donders Institute for Brain, Cognition and Behaviour, Radboud Universiteit, The Netherlands
-   index: 6
- - name: Physikalisch-Technische Bundesanstalt Braunschweig und Berlin, Germany
-   index: 7
-date: 9 March 2025
-bibliography: paper.bib
----
-
-# Summary
-
-Various forms of information can be extracted from electrophysiology data. Of this, phase-amplitude coupling, time delays, and non-sinusoidal waveshape characteristics are of great interest, providing mechanistic insights into physiology and pathology. However, methods commonly used for these analyses possess notable limitations. Recent work has revealed the bispectrum - the Fourier transform of the third order moment - to be a powerful tool for the analysis of electrophysiology data, overcoming many such limitations. Here we present `PyBispectra`, a package for bispectral analyses including phase-amplitude coupling, time delays, and non-sinusoidal waveshape.
-
-# Statement of need
-
-Analysis of phase-amplitude coupling, time delays, and non-sinusoidal waveshape can be applied to various forms of electrophysiology data, including: electroencephalography; magnetoencephalography; local field potentials; electromyography; and electrocardiograhy. These analyses provide important mechanistic insights into bodily functions [@Canolty2010;@Silchenko2010;@Sherman2016], for instance to investigate core nervous system roles like movement and memory, including their perturbation in disease [@deHemptinne2013;@Cole2017;@Bazzigaluppi2018;@Binns2025]. However, traditional analysis methods have critical limitations in terms of robustness, interpretability, and computational demand that hinder their utility. In contrast, the bispectrum measures non-linear signal interactions in a frequency-resolved manner [@Nikias1987], overcoming traditional limitations for the analysis of phase-amplitude coupling [@Zandvoort2021], non-sinusoidal waveshape [@Bartz2019], and time delays [@Nikias1988].
-
[…]
