+++ pyproject.toml
@@ -22,7 +22,7 @@ description = "A Python signal processing package for computing spectral-domain 
 name = "pybispectra"
 readme = "README.md"
 requires-python = ">=3.10"
-version = "1.3.0dev"
+version = "1.2.2"
 
 [project.optional-dependencies]
 dev = ["pybispectra[doc]", "pybispectra[lint]", "pybispectra[test]"]
