+++ src/pybispectra/__init__.py
@@ -1,6 +1,6 @@
 """Initialisation of the PyBispectra package."""
 
-__version__ = "1.2.2"
+__version__ = "1.3.0+dev"
 
 from .cfc import AAC, PAC, PPC
 from .general import Bispectrum, Threenorm
