+++ docs/source/conf.py
@@ -15,7 +15,7 @@ from pybispectra.utils._docs import linkcode_resolve
 project = "PyBispectra"
 copyright = "2023-2025, Thomas S. Binns"
 author = "Thomas S. Binns"
-release = "1.2.2"
+release = "1.3.0dev"
 
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
