### 2025-08-15 e90d40a
[MAINT] Update for new dev version (#92)
