+++ CITATION.cff
@@ -2,7 +2,7 @@ cff-version: 1.2.0
 message: "If you use this software, please cite it as below."
 authors:
 - family-names: "Binns"
-  given-names: "Thomas Samuel"
+  given-names: "Thomas S."
   orcid: "https://orcid.org/0000-0003-0657-0891"
 - family-names: "Pellegrini"
   given-names: "Franziska"
@@ -10,10 +10,16 @@ authors:
 - family-names: "Jurhar"
   given-names: "Tin"
   orcid: "https://orcid.org/0000-0002-8804-2349"
+- family-names: "Nguyen"
+  given-names: "Tien D."
+  orcid: "https://orcid.org/0009-0008-9867-0964"
+- family-names: "Köhler"
+  given-names: "Richard M."
+  orcid: "https://orcid.org/0000-0002-5219-1289"
 - family-names: "Haufe"
   given-names: "Stefan"
   orcid: "https://orcid.org/0000-0003-1470-9195"
-title: "PyBispectra"
+title: "PyBispectra: A toolbox for advanced electrophysiological signal processing using the bispectrum"
 doi: 10.5281/zenodo.8377820
-date-released: 2023-09-23
+date-released: 2025-08-15
 url: "https://github.com/neuromodulation/PyBispectra/tree/main"
