### 2025-08-19 5bbb78d
Bump actions/checkout from 4 to 5 (#96)
Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>
