+++ .github/workflows/unit_tests.yml
@@ -41,7 +41,7 @@ jobs:
       - uses: actions/setup-python@v5
         with:
           python-version: ${{ matrix.python-version }}
-      - uses: actions/checkout@v4
+      - uses: actions/checkout@v5
       - uses: pyvista/setup-headless-display-action@main
         with:
           qt: true
@@ -74,7 +74,7 @@ jobs:
       PYTHONUNBUFFERED: '1'
       PYTHON_VERSION: '3.13'
     steps:
-      - uses: actions/checkout@v4
+      - uses: actions/checkout@v5
       - uses: pyvista/setup-headless-display-action@main
         with:
           qt: true
