### 2025-08-29 290217e
[MAINT] Update toolbox citation
