### 2025-08-29 7401348
[MAINT] Update toolbox citation
