+++ docs/source/index.rst
@@ -25,7 +25,8 @@ implemented to reduce computation times.
 If you use this toolbox in your work, please include the following citation:|br|
 Binns, TS, Pellegrini, F, Jurhar, T, Nguyen, TD, Köhler, RM, & Haufe, S (2025).
 PyBispectra: A toolbox for advanced electrophysiological signal processing using the
-bispectrum. DOI: `10.21105/joss.08504 <https://doi.org/10.21105/joss.08504>`_
+bispectrum. *Journal of Open Source Software*. DOI:
+`10.21105/joss.08504 <https://doi.org/10.21105/joss.08504>`_
 
 .. toctree::
    :maxdepth: 2
