### 2025-08-29 dae6282
[MAINT] Update toolbox citation (#99)
