### 2025-08-29 c971d82
[MAINT] Update toolbox citation (#99)
