### 2025-08-29 f2461b0
[MAINT] Update toolbox citation (#99)
