### 2025-08-29 944d71f
[MAINT] Update toolbox citation (#99)
