### 2025-08-29 95baf46
[MAINT] Update toolbox citation (#99)
