### 2025-08-29 9089dbe
[MAINT] Update toolbox citation (#99)
