### 2025-08-29 56c131f
Add sklearn as dependency
