+++ pyproject.toml
@@ -5,16 +5,14 @@ build-backend = "hatchling.build"
 [project]
 name = "pybispectra"
 version = "1.1.0"
-authors = [
-    { name="Thomas Samuel Binns", email="t.s.binns@outlook.com" }
-]
+authors = [{ name = "Thomas Samuel Binns", email = "t.s.binns@outlook.com" }]
 description = "A Python signal processing package for computing spectral-domain and time-domain interactions using the bispectrum."
 readme = "README.md"
 requires-python = ">=3.10,<3.12"
 classifiers = [
     "Programming Language :: Python :: 3",
     "License :: OSI Approved :: MIT License",
-    "Operating System :: OS Independent"
+    "Operating System :: OS Independent",
 ]
 
 dependencies = [
@@ -25,6 +23,7 @@ dependencies = [
     "pqdm",
     "joblib",
     "matplotlib",
+    "scikit-learn",
 ]
 
 [project.optional-dependencies]
@@ -44,7 +43,7 @@ dev = [
     "notebook",
     "ipykernel",
     "ipywidgets",
-    "ipython"
+    "ipython",
 ]
 
 [project.urls]
@@ -52,11 +51,7 @@ dev = [
 "Bug Tracker" = "https://github.com/braindatalab/PyBispectra/issues"
 
 [tool.coverage.run]
-omit = [
-    "tests/*",
-    "__init__.py",
-    "src/pybispectra/utils/_docs.py"
-]
+omit = ["tests/*", "__init__.py", "src/pybispectra/utils/_docs.py"]
 
 [tool.codespell]
 skip = "./docs/build,./docs/source/auto_examples,./docs/source/refs.bib"
