### 2025-09-02 c42f99c
Return to windows-latest
