+++ .github/workflows/unit_tests.yml
@@ -29,7 +29,7 @@ jobs:
             python-version: "3.13"
           - os: macos-14  # arm64
             python-version: "3.13"
-          - os: windows-2022
+          - os: windows-latest
             python-version: "3.13"
     env:
       TZ: Europe/Berlin
